+++ src/bin_node/node_shared_arg.ml
@@ -51,6 +51,7 @@ type t = {
   rpc_listen_addrs : string list;
   private_mode : bool;
   disable_mempool : bool;
+  disable_mempool_precheck : bool;
   enable_testchain : bool;
   cors_origins : string list;
   cors_headers : string list;
@@ -137,10 +138,10 @@ let load_net_config = function
 let wrap data_dir config_file network connections max_download_speed
     max_upload_speed binary_chunks_size peer_table_size listen_addr
     advertised_net_port discovery_addr peers no_bootstrap_peers
-    bootstrap_threshold private_mode disable_mempool enable_testchain
-    expected_pow rpc_listen_addrs rpc_tls cors_origins cors_headers log_output
-    history_mode synchronisation_threshold latency disable_config_validation
-    allow_all_rpc =
+    bootstrap_threshold private_mode disable_mempool disable_mempool_precheck
+    enable_testchain expected_pow rpc_listen_addrs rpc_tls cors_origins
+    cors_headers log_output history_mode synchronisation_threshold latency
+    disable_config_validation allow_all_rpc =
   let actual_data_dir =
     Option.value ~default:Node_config_file.default_data_dir data_dir
   in
@@ -170,6 +171,7 @@ let wrap data_dir config_file network connections max_download_speed
     rpc_listen_addrs;
     private_mode;
     disable_mempool;
+    disable_mempool_precheck;
     enable_testchain;
     cors_origins;
     cors_headers;
@@ -485,6 +487,17 @@ module Term = struct
     in
     Arg.(value & flag & info ~docs ~doc ["disable-mempool"])
 
+  let disable_mempool_precheck =
+    let doc =
+      "If set to [true], the node's prevalidator will fully execute operations \
+       before gossiping valid operations over the network. Default value is \
+       [false], in which case the node's prevalidator only performs a fast \
+       check over operations before gossiping them. If set to [true], this \
+       option can slow down your node and should be used for testing or \
+       debugging purposes."
+    in
+    Arg.(value & flag & info ~docs ~doc ["disable-mempool-precheck"])
+
   let enable_testchain =
     let doc =
       "If set to [true], the node will spawn a testchain during the protocol's \
@@ -574,10 +587,10 @@ module Term = struct
     $ max_download_speed $ max_upload_speed $ binary_chunks_size
     $ peer_table_size $ listen_addr $ advertised_net_port $ discovery_addr
     $ peers $ no_bootstrap_peers $ bootstrap_threshold $ private_mode
-    $ disable_mempool $ enable_testchain $ expected_pow $ rpc_listen_addrs
-    $ rpc_tls $ cors_origins $ cors_headers $ log_output $ history_mode
-    $ synchronisation_threshold $ latency $ disable_config_validation
-    $ allow_all_rpc
+    $ disable_mempool $ disable_mempool_precheck $ enable_testchain
+    $ expected_pow $ rpc_listen_addrs $ rpc_tls $ cors_origins $ cors_headers
+    $ log_output $ history_mode $ synchronisation_threshold $ latency
+    $ disable_config_validation $ allow_all_rpc
 end
 
 let read_config_file args =
@@ -678,6 +691,7 @@ let read_and_patch_config_file ?(may_override_network = false)
     private_mode;
     discovery_addr;
     disable_mempool;
+    disable_mempool_precheck;
     enable_testchain;
     rpc_listen_addrs;
     rpc_tls;
@@ -815,6 +829,7 @@ let read_and_patch_config_file ?(may_override_network = false)
     ~allow_all_rpc
     ~private_mode
     ~disable_mempool
+    ~disable_mempool_precheck
     ~enable_testchain
     ~cors_origins
     ~cors_headers